buildcompare-sa/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
requests==2.31.0
numpy
python-dotenv
# New dependencies
httpx
beautifulsoup4
Pillow
pytesseract
firebase-admin
groq
# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
